# Reduce allocations in fixBound